+++ src/main/java/dev/yatloaf/modkrowd/cubekrowd/message/KickedMessage.java
@@ -9,15 +9,15 @@ import dev.yatloaf.modkrowd.util.text.StyledStringReader;
 public record KickedMessage(Subserver subserver, String reason, boolean isReal) implements Message {
     public static final KickedMessage FAILURE = new KickedMessage(Subservers.NONE, "", false);
 
-    public static final StyledString CAG2 = StyledString.fromString("Kicked whilst connecting to ", CKColor.RED.style);
+    public static final StyledString PREFIX = StyledString.fromString("Unable to connect to ", CKColor.RED.style);
 
     public static KickedMessage readFast(StyledStringReader source) {
-        if (!source.skipIfNext(CAG2)) return FAILURE;
+        if (!source.skipIfNext(PREFIX)) return FAILURE;
 
         Subserver subserver = Subservers.fromId(source.readUntil(":").toUnstyledString());
         if (!subserver.isReal) return FAILURE;
 
-        source.skipUntilAfter(":");
+        source.skipUntilAfter(": ");
         String reason = source.readAll().toUnstyledString();
 
         return new KickedMessage(subserver, reason, true);