+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Migrated to mojmap.
 - Rewrote lots of code to be more maintainable.
+- Updated parsing in line with CubeKrowd.
 
 ### Fixed